fix(web/src/boot/axios.js): Add setting of Authorization header in interceptor of Axios Instance
Thank the logic inside that interceptor, it occurs the refresh of access token while the refresh token is valid. But, this refreshing is not happening in the intercepted request, what it is resulting in a unauthorized status always that the access token is invalid at the moment that the request is intercepted. After this commit, this problem will no longer exist.