fix(server/events/serializers.py/EventSerializer/validate/): Remove extra "return data"
Because of the "return data" present at the end of the instance if statement, it is possible to set the start period to before the current date and set the end period to before the start period. This issue occurs because "return data" prevents reaching the if statements that prevent these bad edits to the start and end period. After this commit, this "return data" will be removed.